Make sure to use FD -1 for anon mmap to make sparcs and AIX happy
when compressing to 32-bit pointers, we pass 64-bit arguments around.  Make sure
to expect 64-bits on the receiving end or else we have "issues" on big endian
systems like POWER and SPARC.


git-svn-id: https://llvm.org/svn/llvm-project/poolalloc/trunk@20428 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/FL2Allocator/PoolAllocator.h
@@ -205,9 +205,9 @@ extern "C" {
   void *poolinit_pc(PoolTy<CompressedPoolTraits> *Pool, unsigned NodeSize,
                     unsigned ObjAlignment);
   void pooldestroy_pc(PoolTy<CompressedPoolTraits> *Pool);
-  unsigned long poolalloc_pc(PoolTy<CompressedPoolTraits> *Pool,
-                             unsigned NumBytes);
-  void poolfree_pc(PoolTy<CompressedPoolTraits> *Pool, unsigned long Node);
+  unsigned long long poolalloc_pc(PoolTy<CompressedPoolTraits> *Pool,
+                                  unsigned NumBytes);
+  void poolfree_pc(PoolTy<CompressedPoolTraits> *Pool, unsigned long long Node);
   //void *poolmemalign_pc(PoolTy *Pool, unsigned Alignment, unsigned NumBytes);
 }
